feat: add machine geofence query route
