+++ util/locationCheck.ts
@@ -1,4 +1,5 @@
-import {  GroupGeoFenceJSON } from '../models/MachineGroupModel';
+import { MachineGroupDB } from '../models';
+import { GroupGeoFence, GroupGeoFenceJSON, MachineGroup, MachineGroupMachine, ShopLocation } from '../models/MachineGroupModel';
 
 export type Location = {
     lat: number,
@@ -42,6 +43,31 @@ const getDistanceBetweenPoints = (location1:Location, location2:Location) => {
         * Math.sin(Δλ / 2) * Math.sin(Δλ / 2);
     const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
     return R * c;
+};
+export const isMachineRequiresLocation = async (machineId: string): Promise<[boolean, GroupGeoFenceJSON[]]> => {
+    const machineGroupMachine = await MachineGroupDB.findOne({ where: { data: machineId, type: 'MACHINE' } }).then((machineGroup) => machineGroup?.toJSON()) as MachineGroupMachine;
+    if (machineGroupMachine){
+        const machineGroup = await MachineGroupDB.findOne({ where: { id: machineGroupMachine.sk, type: 'GROUP' } }).then((machineGroup) => machineGroup?.toJSON()) as MachineGroup;
+        if (!machineGroup){
+            return [false,[]];
+        }
+        const location = await MachineGroupDB.findOne({ where: { id: machineGroup.sk, type: 'LOCATION' } }).then((machineGroup) => machineGroup?.toJSON()) as ShopLocation;
+        const geoFences = await MachineGroupDB.findAll({ where: { sk: [location?.id, machineGroup.id], type: 'GEOFENCE' } })
+            .then((machineGroups) => machineGroups.map((geoFence) => {
+                const geoFenceObj = geoFence.toJSON() as GroupGeoFence;
+                return {
+                    ...geoFenceObj,
+                    data:JSON.parse(geoFenceObj.data as string) as GeoFence,
+                };
+            })) as GroupGeoFenceJSON[];
+        if (geoFences.length == 0){
+            return [false,[]];
+        }
+        else {
+            return [true, geoFences];
+        }
+    }
+    return [false,[]];
 };
 
 const toRadians = (degrees:number) => degrees * Math.PI / 180;
